Add files via upload
According to the data, the demographic is predominantly Male between the ages of 15-25. Males make up ~84% of the demographic while females are ~14% and other is >~2%. And among that group of males, the older customers are more likely to purchase additional add ins. This is not surprising if you take into consideration income/employment.